requirements.txt
streamlit
pillow
pillow-heif
requests
python-dotenv
PyGithub